于志斌水果商店项目第21次提交 修复了演练时出现的bug 扩展：管理员修改时如果结果为负数要提示

+++ src/manager/dao/impl/FruitDaoImpl.java
@@ -41,7 +41,7 @@ public class FruitDaoImpl implements FruitDao {
     public boolean buyFruit(String name, String amount) {
         Fruit fruit = getByName(name);
         String[] split = fruit.toTxt().split(",");
-        double quantity = Integer.parseInt(split[3]);
+        double quantity = Double.parseDouble(split[3]);
         quantity = quantity - Double.parseDouble(amount);
         if (quantity >= 0) {
             String result = Double.toString(quantity);
